Changing how to use can? to check many subjects
During the test to permit pass an Object as a hash I needed to change how call can? to mantain the functionality to pass any object inclusive an instance of Array as a single subject.
To assure that funciontality, now we need to pass a hash containing a key :any and the array of subjects as his value.
The array deals with the same objects that an usual can? does... and obviously, that hash is only necessary when checking multiple subjects.