Refactoring of method can? to improve legibility
The method can? became a bit messy and repetitive dealing chosing paths almost identical to single and multiple subjects, when only the difference was iterate on the array of subjects.
So, I created the method extract_subject that independently if given a single or multiple subjects, it will returns an array of subjects. So, for can? does  not matter anymore if subject is single or multiple, it will deal on the same way.