Split out separate adapters for AR 3 and AR 4
Detect the AR version and load different adapters
approrpriately. Shared code (which is most of it) is still in
`ActiveRecordAdapter`, though that's been converted into a module so
that both the adapters could remain direct subclasses of
AbstractAdapter (not sure if that was a good idea or not, but it seems
to be working).