Mongoid Adapter: Nested hash conditions
Mongoid adapter now handles nested hash conditions
For eg.
If you have models related in the following manner:
A belongs_to B belongs_to C

and you want to define abilities on A with a hash condition
that concerns fields in C, this commit enables you to do so

Prior to this, can? would work with such nested conditions
but accessible_by would not.

This fixes accessible_by
'