Shorten when success, detail when fail
Previously I was only doing the short version, then I went to all
detailed versions when the object was not an instance_of?(Class) but
when it succeeds we don't need to see those details. If you watch it
fail first then you can see the details for the resource, then when it
passes those details can be hidden. They are always exposed for the
failure case though.