Allow disabling ActiveRecord eager loading
While #includes is useful to help avoid N+1 query problems, it can
cause challenges in other ways. In particular, complex queries that
involve large objects at the root of their chain can end up including
that large object for every row of the selected model.

As an example, imagine a blog post with many comments, each with an
author. We'd like to restrict access to the authors that have comments
on blog posts visible to the user. Your permission might look like:

```
can :read, User, {
  comment: { blog: { accessible_blogs: { user_id: @current_user.id } } }
}
```

Then you can do User.accessible_by(current_ability). That query will
give you a list of all the authors, but will also include the text of
the blog and comment. That becomes an issue for large blog posts with
many comments.

Using joins instead of includes, in this case, runs a query that only
selects User attributes.

Issue cancancan#60 has some more information on an application of this
fix.