Extract resource-loading logic from ControllerResource into another class.
I was reading the source of `ControllerResource` to try to understand
why an `authorize_resource` macro in our controller was authorizing a
different resource than the one we expected. I found it quite hard to
see what was going on, because all the logic for `load_resource` is
mixed into the same class as that for `authorize_resource`.

This commit extracts all the loading logic into a new class called
ControllerResource::Loader, and refactors some internal APIs in the
process to support the interaction between Loader and
ControllerResource. These refactorings include:

* Extracing values that are a matter of definition (e.g. the options to
  `before_filter` and the lists of types of actions) into named
  constants.

* Pulling some repeated expressions into their own methods, e.g.
  `controller_action` and `collection_name`.

* Inlining some methods that are only called from one place, e.g.
  `initial_attributes`, `id_param_key`, `resource_class_with_parent`,
  `params_methods`, `namespace`.

* Eliminate some overly long expressions, and reduce complexity of a few
  methods by flattening conditional blocks.

* Using `__send__` rather than `send` to avoid accidentally calling
  user-defined domain methods called `send`.