Minor ability cleanup.
decendants in the alternative_subjects methods will never not be an
empty array. We can remove this statement.

Make rules an attr_reader and add a bit of documentations as to why it
needs to be protected, as this is one of the -few- cases in ruby where
it makes sense for it to be so.