Fix attempt to #constantize a symbolic name
Related with 146, 6c497b8.

I run into this problem when implementing a successive version of a resource.
I was using something like next in controller:

```ruby
class V2::InvoiceController < V2::BaseController
  load_and_authorize_resource :invoice
end
```

Pretty similar to what docs suggests. But regardless of existing logic that try to compose namespaced resource. It fails due to passing a name of the resource as `:invoice` symbols. And rescuing with a not namespaced resource.

```ruby
   def namespaced_name
      [namespace, name.camelize].flatten.map(&:camelize).join('::').singularize.constantize
    rescue NameError
      name
    end
```

So instead of `V2::Invoice` it returns an `Invoice`.

This PR is about fixing that problem and get rid of `rescue` from almost base error class.