Fix issue with ActiveRecord initialization
Fixes #342

ActiveRecord was being loaded when the model adapter was loaded. This is
too early in the initialization process, as it doesn't allow all of the
Rails initializers to be run first. This patch ensures that the model
adapter code requiring ActiveRecord is not run until after ActiveRecord
has been loaded.