Generate inner queries instead of join+distinct
When applying filters with related entities, CanCan used to generate a LEFT JOIN plus
a DISTINCT clause to avoid repeated records. This is problematic for a number of reasons:

1. Postgres is picky about setting ORDER BY with a column not on the DISTINCT clause, because
   it applies the ORDER after the DISTINCT so it cannot 'see' the referenced column.
2. Groups and Counts become more difficult to do because `foo.distinct.count` is not the same
   as `foo.count`.

The way to avoid this problem is to use inner queries that don't
multiply the number of returned records. We do this by running the
query we were previously using inside a WHERE IN block.