Add support for using an empty array as a condition
I want to define an ability check that only applies if a `has_many` relation returns an empty array. The use case is I want to only allow deleting a record if it has no associated records for a given association.

This PR implements that functionality. Note that it only works for `can?` checks, not `accessible_by`; there's a comment explaining why. I'd like to add a warning to https://github.com/CanCanCommunity/cancancan/wiki/defining-abilities that explains this too if this is approved.