I want to define an ability check that only applies if a `has_many` relation returns an empty array. The use case is I want to only allow deleting a record if it has no associated records for a given association.
In ActiveRecord you can do this with syntax like this:

> Article.joins(:comments).where(comments: { id: nil })

This generates the SQL:

> SELECT "articles".* FROM "articles" INNER JOIN "comments" ON "comments"."article_id" = "articles"."id" WHERE "comments"."id" IS NULL

Which returns an empty array.

This PR makes cancan support the same style of query. So you can do:

> @ability.can :destroy, Article, comments: { id: nil }

This will allow you to `:destroy` an `Article` that has no comments, but not an article that has comments.